requirements.txt
selenium==4.20.0
python-dotenv==1.0.1
